WhisperTRT/requirements.txt
# Core dependencies
numpy>=1.24.0
psutil>=5.9.0
python-dotenv>=1.0.0
pandas>=2.0.0

# PyTorch (should already be installed via JetPack)
# Uncomment if not installed:
# torch>=2.0.0
# torchaudio>=2.0.0

# Whisper dependencies (install before WhisperTRT)
openai-whisper>=20230314
tiktoken>=0.3.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
ffmpeg-python>=0.2.0

# faster-whisper (alternative to WhisperTRT)
faster-whisper>=0.10.0

# Diarization (speaker identification)
pyannote.audio>=3.0.0
torch-audiomentations>=0.11.0

# Additional utilities
tqdm>=4.65.0

# Note: After installing these dependencies, install WhisperTRT separately with:
# cd /path/to/WhisperTRT && pip3 install -e .